# SwaggerOffLine.yaml
openapi: 3.1.0
info:
  title: Auto Escola 3ESR
  description: "API REST para gerenciamento de instrutores, alunos, usuários e agendamento/cancelamento\
    \ de instruções da auto escola 3ESR."
  contact:
    name: Auto Escola 3ESR
    email: autoescola3esr@email.com.br
  license:
    name: Apache 2.0
    url: http://springdoc.org
  version: v1
servers:
- url: http://localhost:8081
  description: Generated server url
...
paths:
...
  /alunos:
    get:
      tags:
      - Alunos
      summary: Listar alunos ativos
      description: "Ordenado por nome (crescente), 10 registros por página."
      operationId: listarAlunos
      parameters:
      - name: page
        in: query
        description: Zero-based page index (0..N)
        required: false
        schema:
          type: integer
          default: 0
          minimum: 0
      - name: size
        in: query
        description: The size of the page to be returned
        required: false
        schema:
          type: integer
          default: 10
          minimum: 1
      - name: sort
        in: query
        description: "Sorting criteria in the format: property,(asc|desc). Default\
          \ sort order is ascending. Multiple sort criteria are supported."
        required: false
        schema:
          type: array
          default:
          - "nome,ASC"
          items:
            type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageDadosListagemAluno"
      security:
      - bearer-key: []
    put:
      tags:
      - Alunos
      summary: Atualizar aluno
      description: "Permite alterar apenas nome, telefone e endereço. E-mail e CPF\
        \ não podem ser alterados."
      operationId: atualizarAluno
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DadosAtualizacaoAluno"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DadosDetalhamentoAluno"
      security:
      - bearer-key: []
    post:
      tags:
      - Alunos
      summary: Cadastrar aluno
      description: Disponível apenas para administradores.
      operationId: cadastrarAluno
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DadosCadastroAluno"
        required: true
      responses:
        "201":
          description: Aluno cadastrado
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DadosDetalhamentoAluno"
        "400":
          description: Dados inválidos
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DadosDetalhamentoAluno"
      security:
      - bearer-key: []
...
  /instrucoes:
    get:
      tags:
      - Instruções
      summary: Listar instruções
      description: "Ordenado por data/hora, 10 registros por página."
      operationId: listarInstrucoes
      parameters:
      - name: page
        in: query
        description: Zero-based page index (0..N)
        required: false
        schema:
          type: integer
          default: 0
          minimum: 0
      - name: size
        in: query
        description: The size of the page to be returned
        required: false
        schema:
          type: integer
          default: 10
          minimum: 1
      - name: sort
        in: query
        description: "Sorting criteria in the format: property,(asc|desc). Default\
          \ sort order is ascending. Multiple sort criteria are supported."
        required: false
        schema:
          type: array
          default:
          - "dataHora,ASC"
          items:
            type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageDadosDetalhamentoAgendamento"
      security:
      - bearer-key: []
    post:
      tags:
      - Instruções
      summary: Agendar instrução
      description: "Formato da data/hora: dd/MM/yyyy - HH:mm. Se o instrutor não for\
        \ informado, um instrutor disponível (opcionalmente da especialidade informada)\
        \ é escolhido aleatoriamente."
      operationId: agendarInstrucao
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DadosAgendamento"
        required: true
      responses:
        "201":
          description: Instrução agendada
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DadosDetalhamentoAgendamento"
        "400":
          description: Dados inválidos ou regra de negócio violada
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DadosDetalhamentoAgendamento"
        "404":
          description: Aluno ou instrutor não encontrado
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DadosDetalhamentoAgendamento"
      security:
      - bearer-key: []
  /instrucoes/{id}/cancelamento:
    post:
      tags:
      - Instruções
      summary: Cancelar instrução
      description: "Motivo obrigatório (ALUNO_DESISTIU, INSTRUTOR_CANCELOU ou OUTROS).\
        \ Somente com antecedência mínima de 24 horas."
      operationId: cancelarInstrucao
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DadosCancelamento"
        required: true
      responses:
        "200":
          description: Instrução cancelada
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DadosDetalhamentoAgendamento"
        "400":
          description: Motivo não informado ou regra de negócio violada
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DadosDetalhamentoAgendamento"
        "404":
          description: Instrução não encontrada
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DadosDetalhamentoAgendamento"
      security:
      - bearer-key: []
...
  /instrucoes/{id}:
    get:
      tags:
      - Instruções
      summary: Detalhar instrução
      operationId: detalharInstrucao
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Instrução não encontrada
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DadosDetalhamentoAgendamento"
      security:
      - bearer-key: []
...
  /enderecos/cep/{cep}:
    get:
      tags:
      - Endereços
      summary: Consultar endereço por CEP
      description: Consome o WebService externo ViaCEP (https://viacep.com.br). Útil
        para preencher automaticamente o endereço no cadastro de alunos e instrutores.
      operationId: consultarCep
      parameters:
      - name: cep
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Endereço encontrado
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DadosConsultaCep"
        "404":
          description: CEP inválido ou inexistente
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DadosConsultaCep"
        "503":
          description: Serviço ViaCEP indisponível
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DadosConsultaCep"
      security:
      - bearer-key: []
  /alunos/{id}:
    get:
      tags:
      - Alunos
      summary: Detalhar aluno
      operationId: detalharAluno
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Aluno encontrado
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DadosDetalhamentoAluno"
        "404":
          description: Aluno não encontrado
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DadosDetalhamentoAluno"
      security:
      - bearer-key: []
    delete:
      tags:
      - Alunos
      summary: Excluir aluno
      description: "Exclusão lógica: o aluno é marcado como inativo."
      operationId: excluirAluno
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: Aluno inativado
      security:
      - bearer-key: []
components:
  schemas:
...
    DadosEndereco:
      type: object
      properties:
        logradouro:
          type: string
          minLength: 1
        numero:
          type: string
        complemento:
          type: string
        bairro:
          type: string
          minLength: 1
        cidade:
          type: string
          minLength: 1
        uf:
          type: string
          minLength: 1
          pattern: "[A-Z]{2}"
        cep:
          type: string
          minLength: 1
          pattern: "[0-9]{5}-[0-9]{3}"
      required:
      - bairro
      - cep
      - cidade
      - logradouro
      - uf
...
    DadosAtualizacaoAluno:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nome:
          type: string
        telefone:
          type: string
        endereco:
          $ref: "#/components/schemas/DadosEndereco"
        email:
          type: string
        cpf:
          type: string
      required:
      - id
    DadosDetalhamentoAluno:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nome:
          type: string
        email:
          type: string
        telefone:
          type: string
        cpf:
          type: string
        endereco:
          $ref: "#/components/schemas/DadosEndereco"
        ativo:
          type: boolean
...
    DadosAgendamento:
      type: object
      properties:
        id_aluno:
          type: integer
          format: int64
        id_instrutor:
          type: integer
          format: int64
        especialidade:
          type: string
          enum:
          - MOTOS
          - CARROS
          - VANS
          - CAMINHOES
        data_hora:
          type: string
          format: date-time
      required:
      - data_hora
      - id_aluno
    DadosDetalhamentoAgendamento:
      type: object
      properties:
        id:
          type: integer
          format: int64
        id_aluno:
          type: integer
          format: int64
        nome_aluno:
          type: string
        id_instrutor:
          type: integer
          format: int64
        nome_instrutor:
          type: string
        especialidade:
          type: string
          enum:
          - MOTOS
          - CARROS
          - VANS
          - CAMINHOES
        data_hora:
          type: string
          format: date-time
        cancelada:
          type: boolean
        motivo_cancelamento:
          type: string
          enum:
          - ALUNO_DESISTIU
          - INSTRUTOR_CANCELOU
          - OUTROS
    DadosCancelamento:
      type: object
      properties:
        motivo:
          type: string
          enum:
          - ALUNO_DESISTIU
          - INSTRUTOR_CANCELOU
          - OUTROS
      required:
      - motivo
    DadosCadastroAluno:
      type: object
      properties:
        nome:
          type: string
          minLength: 1
        email:
          type: string
          format: email
          minLength: 1
        telefone:
          type: string
          minLength: 1
        cpf:
          type: string
          minLength: 1
          pattern: "\\d{11}"
        endereco:
          $ref: "#/components/schemas/DadosEndereco"
      required:
      - cpf
      - email
      - endereco
      - nome
      - telefone
...
    PageableObject:
      type: object
      properties:
        paged:
          type: boolean
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        offset:
          type: integer
          format: int64
        sort:
          $ref: "#/components/schemas/SortObject"
        unpaged:
          type: boolean
    SortObject:
      type: object
      properties:
        sorted:
          type: boolean
        empty:
          type: boolean
        unsorted:
          type: boolean
...
    PageDadosDetalhamentoAgendamento:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: "#/components/schemas/DadosDetalhamentoAgendamento"
        number:
          type: integer
          format: int32
        sort:
          $ref: "#/components/schemas/SortObject"
        pageable:
          $ref: "#/components/schemas/PageableObject"
        numberOfElements:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean
        empty:
          type: boolean
    DadosConsultaCep:
      type: object
      properties:
        cep:
          type: string
        logradouro:
          type: string
        complemento:
          type: string
        bairro:
          type: string
        cidade:
          type: string
        uf:
          type: string
    DadosListagemAluno:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nome:
          type: string
        email:
          type: string
        cpf:
          type: string
    PageDadosListagemAluno:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: "#/components/schemas/DadosListagemAluno"
        number:
          type: integer
          format: int32
        sort:
          $ref: "#/components/schemas/SortObject"
        pageable:
          $ref: "#/components/schemas/PageableObject"
        numberOfElements:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean
        empty:
          type: boolean
  securitySchemes:
    bearer-key:
      type: http
      description: Informe o token JWT obtido em POST /login
      scheme: bearer
      bearerFormat: JWT